refactor(pokemon): centralize Champions Reg M-A legality into one bundle + registry
Move M-A's curated legality data (legal species/items/moves/abilities, mega stones/abilities/stats) out of format-legality.ts and stats-calculator.ts into a new leaf module champions-reg-ma.ts exporting REG_MA_BUNDLE. Replace the hardcoded `formatId === MA_ID` checks in the five getLegal* dispatch fns with a CHAMPIONS_LEGALITY_BY_ID registry, so adding a regulation becomes one data file + one registry line instead of editing dispatch in 5+ spots.

Pure data move — M-A behavior unchanged (code-reviewer verdict: behavior-preserving across all 5 dispatch fns; reg-ma.ts is a true leaf, no import cycle).
